feat: change method name to updateMaturityDate
