+++ sdk/src/core/Constants.ts
@@ -274,7 +274,7 @@ export const UPDATE_CONFIG_VERSION_GAS = 9000000;
 export const UPDATE_CONFIG_GAS = 9000000;
 export const UPDATE_RESOLVER_GAS = 9000000;
 
-export const SET_MATURITY_DATE_GAS = 7000000;
+export const UPDATE_MATURITY_DATE_GAS = 7000000;
 
 export const _PARTITION_ID_1 =
   '0x0000000000000000000000000000000000000000000000000000000000000001';
