+++ sdk/src/port/in/request/index.ts
@@ -257,7 +257,7 @@ import GetConfigInfoRequest from './GetConfigInfoRequest.js';
 import UpdateConfigRequest from './UpdateConfigRequest.js';
 import UpdateConfigVersionRequest from './UpdateConfigVersionRequest.js';
 import UpdateResolverRequest from './UpdateResolverRequest.js';
-import SetMaturityDateRequest from './SetMaturityDateRequest.js';
+import UpdateMaturityDateRequest from './UpdateMaturityDateRequest.js';
 
 export * from './BaseRequest.js';
 export {
@@ -314,5 +314,5 @@ export {
   UpdateConfigVersionRequest,
   UpdateConfigRequest,
   GetConfigInfoRequest,
-  SetMaturityDateRequest,
+  UpdateMaturityDateRequest,
 };
